requirements.txt
argparse
boto3
click
flask
matplotlib
pandas
pytest
python-dotenv
pyautogui
pyscreeze
pytesseract
Pillow
plyer
PyYAML
requests
simpy
sqlalchemy
